Replace NavigationView with NavigationStack and refactor name changes
