--- SnapSpire/Features/Discover/View/CollectionView.swift
@@ -1,13 +1,0 @@
-import SwiftUI
-
-struct CollectionView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct CollectionView_Previews: PreviewProvider {
-    static var previews: some View {
-        CollectionView()
-    }
-}
