+++ SnapSpire/Resources/Assets.xcassets/Discovery/design.imageset/Contents.json
