+++ SnapSpire/Resources/Assets.xcassets/Discovery/photo.imageset/Contents.json
